auth_test: move running_callback_server fixture into cli/conftest.py
Problem: auth_test.py defined a @pytest.fixture (running_callback_server)
inline. The root CLAUDE.md says "All fixtures must be in conftest.py, not
in individual test files." No other test file under libs/mngr_imbue_cloud
defines fixtures inline.

Fix: extract the fixture into a new libs/mngr_imbue_cloud/imbue/
mngr_imbue_cloud/cli/conftest.py so it is auto-discovered by pytest for
any test under cli/. auth_test.py only retains the small _get helper and
the import of _OAuthCaptureBox needed for the param annotations.
